backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "ProfileForge AI - Academic Profile Management with AI Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.134.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=2.0.9",
    "langchain-community>=0.3.20",
    "pydantic>=2.12.5",
    "email-validator>=2.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "python-multipart>=0.0.19",
]
